+++ src/StkTable/custom-cells/EditableCell/index.ts
@@ -1,0 +1,3 @@
+export { default as EditableCell } from './EditableCell.vue';
+export { createEditableCell } from './createEditableCell';
+export type { CreateEditableCellOptions } from './createEditableCell';